store submissions in player's object and Delegate working
