+++ PartyGame/Services/GameCenterService.swift
@@ -12,7 +12,7 @@ extension UIApplication {
     }
 }
 
-private struct LobbyPacket: Codable {
+struct LobbyPacket: Codable {
     enum PacketType: String, Codable { case chat, ready}
     let type: PacketType
     let senderID: String
@@ -56,9 +56,9 @@ class GameCenterService: NSObject, ObservableObject {
     @Published var timerStart: Date? = nil
     
     var match: GKMatch?
-    private var pendingInvite: GKInvite?
-    private var pendingPlayersToInvite: [GKPlayer]?
-    private var expectedPlayersCount: Int {
+    internal var pendingInvite: GKInvite?
+    internal var pendingPlayersToInvite: [GKPlayer]?
+    internal var expectedPlayersCount: Int {
         if isSinglePlayer { return 1 }
         let ids = Set(([GKLocalPlayer.local] + (match?.players ?? [])).map { $0.gamePlayerID })
         return ids.count
@@ -283,6 +283,39 @@ class GameCenterService: NSObject, ObservableObject {
         return ((gamePlayers.count == playerSubmissions.count && gamePlayers.count != 0) ? true : false)
     }
     
+    func cleanAndStorePlayerSubmissions() {
+        addSubmissionToPlayers()
+        
+        cleanPlayerSubmissions(broadcast: true)
+    }
+    
+    func cleanPlayerSubmissions(broadcast: Bool) {
+        playerSubmissions.removeAll()
+        print("playerSubmissions were cleaned")
+        
+        guard broadcast, let match else { return }
+        
+        do {
+            let payload = ["type": "CleanPlayerSubmissions"]
+            let data = try JSONEncoder().encode(payload)
+            try match.sendData(toAllPlayers: data, with: .reliable)
+            print("📡 Submissões de imagens enviadas.")
+        } catch {
+            print("❌ Erro ao enviar submissões de imagens: \(error)")
+        }
+    }
+    
+    func addSubmissionToPlayers() {
+        for submission in playerSubmissions {
+            let gamePlayerIndex = gamePlayers.firstIndex(where: {$0.player.gamePlayerID == submission.playerID})
+            if let index = gamePlayerIndex {
+                gamePlayers[index].submissions.append(submission)
+                print("player: \(gamePlayers[index].player.displayName) - Submissão adicionada: \(gamePlayers[index].submissions)")
+            }
+            
+        }
+    }
+    
     //MARK: submissão de imagem do jogador para a frase atual
     func addSubmission(playerID: String, phrase: String, image: ImageSubmission) {
         let submission = PlayerSubmission(playerID: playerID, phrase: phrase, imageSubmission: image, votes: 0)
@@ -483,7 +516,7 @@ class GameCenterService: NSObject, ObservableObject {
         }
     }
     
-    private func trySelectPhraseIfReady() {
+    internal func trySelectPhraseIfReady() {
         guard currentPhrase.isEmpty else { return }
         guard let leaderID = phraseLeaderID else { return }
 
@@ -498,7 +531,7 @@ class GameCenterService: NSObject, ObservableObject {
         }
     }
     
-    private func refreshPlayers() {
+    internal func refreshPlayers() {
         var everyone: [GKPlayer] = [GKLocalPlayer.local as GKPlayer]
         if let remotes = match?.players { everyone.append(contentsOf: remotes) }
         DispatchQueue.main.async {
@@ -534,175 +567,4 @@ class GameCenterService: NSObject, ObservableObject {
     
 }
 
-// MARK: - Delegates
-extension GameCenterService: GKMatchmakerViewControllerDelegate, GKMatchDelegate {
-    func matchmakerViewControllerWasCancelled(_ viewController: GKMatchmakerViewController) {
-        viewController.dismiss(animated: true)
-        print("❌ Matchmaking cancelado")
-    }
-    
-    func matchmakerViewController(_ viewController: GKMatchmakerViewController,
-                                  didFailWithError error: Error) {
-        viewController.dismiss(animated: true)
-        print("❌ Erro no matchmaking: \(error.localizedDescription)")
-    }
-    
-    func matchmakerViewController(_ viewController: GKMatchmakerViewController,
-                                  didFind match: GKMatch) {
-        viewController.dismiss(animated: true)
-        self.match = match
-        match.delegate = self
-        print("✅ Match encontrado com \(match.players.count) jogadores")
-        
-        refreshPlayers()
-        let localID = GKLocalPlayer.local.gamePlayerID
-        DispatchQueue.main.async {
-            self.isInMatch = true
-            var map: [String: Bool] = [:]
-            map[localID] = false
-            for p in match.players { map[p.gamePlayerID] = false }
-            self.phrases = []
-            self.readyMap = map
-        }
-        
-        sendMessage("Olá, galera!")
-    }
-    
-    func match(_ match: GKMatch, didReceive data: Data, fromRemotePlayer player: GKPlayer) {
-        
-        guard
-            let dict = try? JSONSerialization.jsonObject(with: data) as? [String: Any],
-            let type = dict["type"] as? String
-        else { return }
-        
-        switch type {
-        case "PhraseLeader":
-            if let leaderID = dict["leaderID"] as? String {
-                DispatchQueue.main.async {
-                    self.phraseLeaderID = leaderID
-                    print("📡 Líder da frase recebido: \(leaderID)")
-                    self.trySelectPhraseIfReady()
-                }
-            }
-        case "SelectedPhrase":
-            if let phrase = dict["currentPhrase"] as? String {
-                DispatchQueue.main.async {
-                    if self.currentPhrase.isEmpty {
-                        self.currentPhrase = phrase
-                        self.isWaitingForPhrase = false
-                        print("📡 Frase selecionada recebida: \(phrase)")
-                    } else {
-                        print("⚠️ Frase já estava definida: \(self.currentPhrase)")
-                    }
-                }
-            }
-        default:
-            break
-        }
-        
-        guard
-            let dict = try? JSONSerialization.jsonObject(with: data) as? [String: Any],
-            let type = dict["type"] as? String
-        else { return }
-        
-        switch type {
-        case "newPhrase":
-            if let phrase = dict["phrase"] as? String {
-                print("frase adicionada Delegate")
-                phrases.append(phrase)
-                self.trySelectPhraseIfReady()
-            }
-            
-        default:
-            break
-        }
-        
-        if let payload = try? JSONDecoder().decode(SubmissionPayload.self, from: data) {
-            switch payload.type {
-            case "newImage":
-                let submission = payload.submission
-                DispatchQueue.main.async {
-                    self.playerSubmissions.append(submission)
-                    print("Nova submissão recebida e adicionada: \(submission)")
-                    print("Printar todos jogadores: \(self.gamePlayers)")
-                }
-            default:
-                break
-            }
-        }
-        
-        if let packet = try? JSONDecoder().decode(LobbyPacket.self, from: data) {
-            switch packet.type {
-            case .chat:
-                if let text = packet.text {
-                    DispatchQueue.main.async {
-                        self.messages.append("\(player.displayName): \(text)")
-                    }
-                }
-            case .ready:
-                let value = packet.ready ?? false
-                DispatchQueue.main.async {
-                    self.readyMap[player.gamePlayerID] = value
-                }
-            }
-        } else if let text = String(data: data, encoding: .utf8) {
-            DispatchQueue.main.async {
-                self.messages.append("\(player.displayName): \(text)")
-            }
-        }
-    }
-    
-    func match(_ match: GKMatch, player: GKPlayer, didChange state: GKPlayerConnectionState) {
-        switch state {
-        case .connected:
-            print("✅ \(player.displayName) conectado")
-            refreshPlayers()
-        case .disconnected:
-            print("❌ \(player.displayName) desconectado")
-            refreshPlayers()
-        default:
-            print("⚠️ Estado desconhecido para \(player.displayName)")
-        }
-    }
-    
-    func match(_ match: GKMatch, didFailWithError error: Error?) {
-        print("❌ Erro no match: \(error?.localizedDescription ?? "desconhecido")")
-        leaveMatch()
-    }
-    
-}
 
-// MARK: - Listener de convites
-extension GameCenterService: GKLocalPlayerListener {
-    // Quando um convite é aceito pelo usuário FORA do app
-    func player(_ player: GKPlayer, didAccept invite: GKInvite) {
-        print("📩 Convite recebido de \(invite.sender.displayName)")
-        
-        // Armazenar o convite para processamento
-        pendingInvite = invite
-        
-        // Se o app estiver ativo, processar imediatamente
-        if UIApplication.shared.applicationState == .active {
-            DispatchQueue.main.asyncAfter(deadline: .now() + 0.5) {
-                self.processPendingInvite()
-            }
-        }
-        // Se o app não estiver ativo, será processado quando se tornar ativo
-    }
-    
-    // Recebimento de solicitação de partida
-    func player(_ player: GKPlayer, didRequestMatchWithRecipients recipientPlayers: [GKPlayer]) {
-        print("📩 Solicitação de partida recebida para \(recipientPlayers.count) jogadores")
-        
-        // Armazenar a solicitação para processamento
-        pendingPlayersToInvite = recipientPlayers
-        
-        // Se o app estiver ativo, processar imediatamente
-        if UIApplication.shared.applicationState == .active {
-            DispatchQueue.main.asyncAfter(deadline: .now() + 0.5) {
-                self.processPendingInvite()
-            }
-        }
-        // Se o app não estiver ativo, será processado quando se tornar ativo
-    }
-}
